+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## main / unreleased
-
+* [FEATURE] Add the `/api/status/buildinfo` endpoint [#2702](https://github.com/grafana/tempo/pull/2702) (@fabrizio-grafana)
 * [FEATURE] New encoding vParquet3 with support for dedicated attribute columns (@mapno, @stoewer) [#2649](https://github.com/grafana/tempo/pull/2649)
 
 ## v2.2.0-rc0 / 2023-07-21